% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/calc.fo.weights.R
\name{calc.fo.weights}
\alias{calc.fo.weights}
\title{Calculate Fictional Observation Weights}
\usage{
calc.fo.weights(formula, label, ratio, data)
}
\arguments{
\item{formula}{lm-style formula}

\item{label}{the term to be used in fictional observations}

\item{ratio}{desired ratio of observed and fictional data}

\item{data}{optional data frame}
}
\value{
Coefficients for fictional observations
}
\description{
Determines the coeffiicients for terms in the fictional observations,
to ensure a given weight relative to the normal observations.
}
\examples{
calc.fo.weights(Sepal.Length ~ Petal.Length + Sepal.Width : Species,
  "Sepal.Width:Species", 5, data=iris)
}

