% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/combine.data.frames.R
\name{combine.data.frames}
\alias{combine.data.frames}
\title{Combine two data frames}
\usage{
combine.data.frames(data1, data2, sep = "", keepfactors = NULL)
}
\arguments{
\item{data1}{data frame}

\item{data2}{data frame}

\item{sep}{a paste separator for creating factor names}

\item{keepfactors}{a list of factor names to translate (or NULL for
all); excluded factors will be given columns set to 0}
}
\value{
Augmented data frame
}
\description{
Add rows as necessary to construct a dataframe with observations
from both of the source data frames.
}
\details{
Translates factors into multiple columns as needed.
}
\examples{
combine.data.frames(iris, data.frame(Sepal.Length=1, Matched.Sepal.Length=-1))
combine.data.frames(iris, data.frame(Species="something else"))
}

