% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/get.formula.terms.R
\name{get.formula.terms}
\alias{get.formula.terms}
\title{Extract the terms from an lm formula}
\usage{
get.formula.terms(formula, all.names = F)
}
\arguments{
\item{formula}{lm-style formula}

\item{all.names}{boolean to extract every unique name}
}
\value{
Vector of the terms in the lm formula
}
\description{
Return a list of terms in an lm formula, in the order they appear
}
\examples{
get.formula.terms(Sepal.Length ~ Petal.Length + Species + Sepal.Width : Species)
}

