% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/get.hyper.terms.R
\name{get.hyper.terms}
\alias{get.hyper.terms}
\title{Get any new hyper terms from formula}
\usage{
get.hyper.terms(formula, sep = "")
}
\arguments{
\item{formula}{Single component of a hierlm formula}

\item{sep}{a paste separator for creating hyper terms}
}
\value{
Vector of c(nonhyper, hyper, is.interaction), if a new one was created;
  or c(left, right, NA) if both terms are given;
  or c() if not an "a - b" expression
}
\description{
If the "a - b" formula requires the creation of a new hyper terms
(one of a or b is .), return them.
Can also be of the form "a > b"
}
\examples{
get.hyper.terms(Sepal.Length ~ Petal.Length + Species + Sepal.Width : Species)
}

