% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/make.factor.names.R
\name{make.factor.names}
\alias{make.factor.names}
\title{Make a factor name}
\usage{
make.factor.names(name, levels, sep = "")
}
\arguments{
\item{name}{the factor to be decomposed}

\item{levels}{data frame with that as a column, or levels to produce}

\item{sep}{a paste separator for creating factor names}
}
\value{
Vector of the factor names
}
\description{
Construct a factor name for each level
}
\examples{
make.factor.names("Species", iris)
make.factor.names("Species", unique(iris$Species))
}

