% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/make.hyper.rows.R
\name{make.hyper.rows}
\alias{make.hyper.rows}
\title{Create fictional observations for a hyper relationship}
\usage{
make.hyper.rows(nonhyper, hyper, weight, data, sep = "")
}
\arguments{
\item{nonhyper}{name of the non-hyper column}

\item{hyper}{name of the hyper column}

\item{weight}{weight to be used in the rows}

\item{data}{a data frame for inspecting levels}

\item{sep}{a paste separator for creating hyper terms}
}
\value{
Fictional observations in a data frame
}
\description{
Return a data frame with fictional observations
}
\examples{
make.hyper.rows("Species", "hyperSpecies", 5, iris)
}

