% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/replace.factor.data.frame.R
\name{replace.factor.data.frame}
\alias{replace.factor.data.frame}
\title{Replace a factor column}
\usage{
replace.factor.data.frame(name, data, iftrue, sep = "")
}
\arguments{
\item{name}{column name}

\item{data}{data frame}

\item{iftrue}{value to enter in column for the indicator}

\item{sep}{paste sep used to construct factor names}
}
\value{
Modified data frame
}
\description{
Constructs indicator columns for a factor variable.
}
\examples{
replace.factor.data.frame("Species", data=iris, 1)
}

