% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/replace.factor.formula.R
\name{replace.factor.formula}
\alias{replace.factor.formula}
\title{Replace a factor term}
\usage{
replace.factor.formula(name, formula, data, sep = "")
}
\arguments{
\item{name}{term name}

\item{formula}{formula}

\item{data}{a data frame to be used to identify factor levels}

\item{sep}{paste sep used to construct factor names}
}
\value{
Modified formula
}
\description{
Constructs indicator terms for a factor variable.
}
\examples{
replace.factor.formula("Species", Sepal.Length ~ Species, iris)
replace.factor.formula("Species", Sepal.Length ~ Petal.Length : Species, iris)
}

