% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/calc.fo.weights.R
\name{calc.fo.weights}
\alias{calc.fo.weights}
\title{Calculate Fictional Observation Weights}
\usage{
calc.fo.weights(ratios, formula, data)
}
\arguments{
\item{ratios}{desired ratio of observed and fictional data}

\item{formula}{lm-style formula}

\item{data}{optional data frame}
}
\value{
Coefficients for fictional observations
}
\description{
Determines the coeffiicients for terms in the fictional observations,
to ensure a given weight relative to the normal observations.
}
\details{
Assumes only one contrast.
}
\examples{
calc.fo.weights(5, Sepal.Length ~ Petal.Length + Sepal.Width : Species, data=iris)
}

